Refactor to use TOML for Configuration
Refactored to use a TOML file for configuration instead of
command line arguments. This change should make it easier
for the end user to configure the program without the need
of remembering arguments. A default config will be shipped
with the build with default values. Also in this commit,
all things relating to configuration have been moved to
their own file for better organization.